in the low-level input routines, make the ascii_digits variables ints instead of size_t's, since they will only hold very small values
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2136 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/proto.h
@@ -531,7 +531,7 @@ int get_translated_kbinput(int kbinput, seq_type *seq
 	, bool reset
 #endif
 	);
-int get_ascii_kbinput(int kbinput, size_t ascii_digits
+int get_ascii_kbinput(int kbinput, int ascii_digits
 #ifndef NANO_SMALL
 	, bool reset
 #endif
